# api-contract.yaml
openapi: 3.0.3
info:
  title: GroceryHack API
  version: 1.0.0
  description: |
    API contract for the GroceryHack application.
    All endpoints require Authorization header with Bearer token unless marked as public.
    All timestamps are ISO 8601 UTC. All monetary values are decimal numbers.
servers:
  - url: /api/v1

# ────────────────────────────────────────────────────────────
# TAGS
# ────────────────────────────────────────────────────────────
...
paths:
...
  /meals:
    get:
      tags: [Meals]
      summary: Get meals for swiping (filtered by user's dietary restrictions)
      description: |
        Returns meals from the shared pool + public user recipes,
        filtered by the user's dietary restrictions,
        excluding meals the user has already swiped.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Swipeable meals with attribution
          content:
            application/json:
              schema:
                type: object
                properties:
                  meals:
                    type: array
                    items:
                      $ref: '#/components/schemas/SwipeableMeal'
  /meals/{meal_id}:
    get:
      tags: [Meals]
      summary: Get full meal detail with attribution
      parameters:
        - name: meal_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Meal detail with attribution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwipeableMeal'
        '404':
          $ref: '#/components/responses/NotFound'
  /meals/{meal_id}/swipe:
    post:
      tags: [Meals]
      summary: Record a swipe (like or skip)
      parameters:
        - name: meal_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [liked]
              properties:
                liked:
                  type: boolean
      responses:
        '200':
          description: Swipe recorded, updated taste profile returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  preference:
                    $ref: '#/components/schemas/UserMealPreference'
                  taste_profile:
                    type: object
                    additionalProperties:
                      type: number
        '409':
          description: Already swiped this meal
...
  /recipes:
    get:
      tags: [UserRecipes]
      summary: Get current user's recipes
      responses:
        '200':
          description: User recipes
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipes:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserRecipe'

    post:
      tags: [UserRecipes]
      summary: Create a recipe manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRecipeCreate'
      responses:
        '201':
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecipe'
        '400':
          $ref: '#/components/responses/BadRequest'
  /recipes/{recipe_id}:
    get:
      tags: [UserRecipes]
      summary: Get a specific recipe
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recipe detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecipe'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [UserRecipes]
      summary: Update a recipe
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRecipeCreate'
      responses:
        '200':
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecipe'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [UserRecipes]
      summary: Delete a recipe
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Recipe deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /recipes/{recipe_id}/stats:
    get:
      tags: [UserRecipes]
      summary: Get match stats for a user's recipe
      description: Only accessible by the recipe owner. Shows how the recipe is performing in the shared pool.
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recipe stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeStats'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          description: Not the recipe owner
  /recipes/{recipe_id}/publish:
    post:
      tags: [UserRecipes]
      summary: Toggle public visibility of a recipe
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_public]
              properties:
                is_public:
                  type: boolean
      responses:
        '200':
          description: Visibility updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecipe'

  # ══════════════════════════════════════
  # DEAL WATCHLIST
  # ══════════════════════════════════════
  /watchlist:
    get:
      tags: [DealWatchlist]
      summary: Get user's watched deals
      responses:
        '200':
          description: Watchlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WatchlistItem'

    post:
      tags: [DealWatchlist]
      summary: Heart a deal (add to watchlist)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deal_id]
              properties:
                deal_id:
                  type: string
                  format: uuid
                  description: The deal being hearted — server extracts product info from the deal
      responses:
        '201':
          description: Added to watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItem'
  /watchlist/{watchlist_id}:
    delete:
      tags: [DealWatchlist]
      summary: Remove from watchlist
      parameters:
        - name: watchlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed

  # ══════════════════════════════════════
  # IMPORTANT ITEMS
  # ══════════════════════════════════════
...
  /share/meal:
    post:
      tags: [Sharing]
      summary: Send a meal to someone ("Cook this for me?" / "I'll make this for you!")
      description: |
        Creates a meal_shares record and sends email/SMS to recipient.
        For cook_for_me: email includes Accept/Decline buttons.
        For make_for_you: informational only (no response needed).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [meal_id, recipient_contact, share_type]
              properties:
                meal_id:
                  type: string
                  format: uuid
                meal_source:
                  type: string
                  enum: [meal, user_recipe]
                  default: meal
                  description: Whether the meal is from the shared pool or user recipes
                recipient_name:
                  type: string
                recipient_contact:
                  type: string
                  description: Email address or phone number
                share_type:
                  type: string
                  enum: [cook_for_me, make_for_you]
                date:
                  type: string
                  format: date
                time:
                  type: string
                  description: "HH:MM format"
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent:
                    type: boolean
                  channel:
                    type: string
                    enum: [email, sms]
                  share_token:
                    type: string
                    description: Token for the meal_shares record (can track status)
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ══════════════════════════════════════
  # SHARED RESPONSES
  # ══════════════════════════════════════
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [error, code, message]
      properties:
        error:
          type: boolean
          enum: [true]
        code:
          type: string
          description: Machine-readable error code
          example: NOT_FOUND
        message:
          type: string
          description: Human-readable message
          example: No stores found near that postal code
...
    Nutrition:
      type: object
      properties:
        calories:
          type: number
        protein_g:
          type: number
        carbs_g:
          type: number
        fat_g:
          type: number
        fiber_g:
          type: number
        sodium_mg:
          type: number
        per_serving:
          type: boolean
          default: true
    Ingredient:
      type: object
      required: [name, quantity, unit]
      properties:
        name:
          type: string
        quantity:
          type: string
        unit:
          type: string
    Meal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        instructions:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        steps:
          type: array
          items:
            type: string
        prep_time_minutes:
          type: integer
        cook_time_minutes:
          type: integer
        servings:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium]
        filter_tags:
          type: array
          items:
            type: string
        taste_tags:
          type: object
        tips:
          type: string
        nutrition:
          $ref: '#/components/schemas/Nutrition'
        created_at:
          type: string
          format: date-time
    MealAttribution:
      type: object
      description: Attribution fields for public user recipes in the shared pool
      properties:
        shared_by_name:
          type: string
          nullable: true
          description: Uploader's display_name. Null for system-generated meals.
        shared_by_initials:
          type: string
          nullable: true
          description: First letter(s) of display_name, max 2 chars. Null for system meals.
        weekly_match_count:
          type: integer
          description: Count of users who swiped right on this meal this week.
    SwipeableMeal:
      description: Meal with attribution fields, used in landing page and swipe deck
      allOf:
        - $ref: '#/components/schemas/Meal'
        - $ref: '#/components/schemas/MealAttribution'
    RecipeStats:
      type: object
      description: Performance stats for a user's public recipe
      properties:
        recipe_id:
          type: string
          format: uuid
        is_public:
          type: boolean
        weekly_match_count:
          type: integer
          description: Users who swiped right this week
        total_match_count:
          type: integer
          description: All-time right swipes
...
    UserRecipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        instructions:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        steps:
          type: array
          items:
            type: string
        prep_time_minutes:
          type: integer
        cook_time_minutes:
          type: integer
        servings:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium]
        dietary_tags:
          type: array
          items:
            type: string
        taste_tags:
          type: object
        tips:
          type: string
        ingredient_keywords:
          type: array
          items:
            type: string
        cost_drivers:
          type: array
          items:
            type: string
        nutrition:
          $ref: '#/components/schemas/Nutrition'
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserRecipeCreate:
      type: object
      required: [name, ingredients]
      properties:
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        instructions:
          type: string
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        steps:
          type: array
          items:
            type: string
        prep_time_minutes:
          type: integer
        cook_time_minutes:
          type: integer
        servings:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium]
        dietary_tags:
          type: array
          items:
            type: string
        tips:
          type: string
        nutrition:
          $ref: '#/components/schemas/Nutrition'
        is_public:
          type: boolean
          default: false
    UserMealPreference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        meal_id:
          type: string
          format: uuid
        liked:
          type: boolean
        swiped_at:
          type: string
          format: date-time
    WatchlistItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        item_keyword:
          type: string
        product_name:
          type: string
        category:
          type: string
        subcategory:
          type: string
        product_metadata:
          type: object
        price_tier:
          type: string
          enum: [staple, premium, luxury]
        benchmark_price:
          type: number
        benchmark_unit:
          type: string
        store_brand_id:
          type: string
          format: uuid
          nullable: true
        store_brand_name:
          type: string
          nullable: true
          description: Brand name resolved from store_brand_id for display convenience
        created_at:
          type: string
          format: date-time
...
security:
  - BearerAuth: []